refactor: GameResult 열거형 패키지 이동
GameResult 열거형 클래스를 어플리케이션 상수 클래스를 모아놓은 constant 패키지로 이동함.
- FinalResult에 생긴 변경사항(GameResult 클래스 import)도 함께 반영함.
